refactor: SVG 파일 이동
- core/assets -> entities/card/brand/assets
- 카드프리뷰에서 브랜드로 svg 맵 이동
