+++ src/entities/card/bank/bank.ts
