+++ src/features/cardPreview/CardPreview.tsx
@@ -1,16 +1,10 @@
 import styles from './CardPreview.module.css';
-import bankStyles from '@/entities/card/bank.module.css';
+import bankStyles from '@/entities/card/bank/bank.module.css';
 
-import AmericanExpressSvg from '@/core/assets/AmericanExpress.svg?react';
-import DinersSvg from '@/core/assets/Diners.svg?react';
-import UnionPaySvg from '@/core/assets/UnionPay.svg?react';
-import MastercardSvg from '@/core/assets/Mastercard.svg?react';
-import VisaSvg from '@/core/assets/Visa.svg?react';
-
-import { getBrand, RULES, type Brand } from '@/entities/card/brand';
+import { getBrand, RULES, type Brand } from '@/entities/card/brand/brand';
 import type { ExpiryDate } from '@/entities/card/expiryDate';
-import type { ReactNode } from 'react';
-import type { Bank } from '@/entities/card/bank';
+import type { Bank } from '@/entities/card/bank/bank';
+import { BRAND_SVG_MAP } from '@/entities/card/brand/brandSvgMap';
 
 export interface CardInfo {
   cardNumbers: string[];
@@ -21,15 +15,6 @@ interface CardPreviewProps {
   info: CardInfo;
 }
 
-const BRAND_SVG_MAP: Record<Brand, ReactNode> = {
-  VISA: <VisaSvg />,
-  MASTERCARD: <MastercardSvg />,
-  AMEX: <AmericanExpressSvg />,
-  DINERS: <DinersSvg />,
-  UNIONPAY: <UnionPaySvg />,
-  UNKNOWN: undefined,
-};
-
 const STAR = '●';
 
 const CardNumber = ({ cardNumbers, brand }: { cardNumbers: string[]; brand: Brand }) => {
@@ -37,7 +22,7 @@ const CardNumber = ({ cardNumbers, brand }: { cardNumbers: string[]; brand: Bran
   return (
     <>
       {INPUT_FORMAT.map((size, i) => (
-        <span key={`${brand}-card-number-${size}-${i}`} className={styles.cardNumberSection}>
+        <span key={`card-number-${i}`} className={styles.cardNumberSection}>
           {i >= 2 ? STAR.repeat(cardNumbers[i]?.length ?? 0) : (cardNumbers[i] ?? '')}
         </span>
       ))}
